requirements.txt
requests
python-dotenv 
beautifulsoup4
mysql-connector-python
openai
python-dotenv